Almost everything works now :)))
Everything does actually work just that the lobby player doesn't get its parent and transform updated on the client. The lobby player hook gets called even on the client, however the method that subscribes to the callback doesn't get called. Not that it should matter since the lobby player has a network transform but...who knows. But tHIS IS HUGE!!!